chore: move code according to comments
