+++ packages/fx-core/tests/component/driver/aad/update.test.ts
@@ -6,10 +6,7 @@ import * as sinon from "sinon";
 import mockedEnv, { RestoreFn } from "mocked-env";
 import chai from "chai";
 import chaiAsPromised from "chai-as-promised";
-import {
-  UpdateAadAppDriver,
-  getQuestionForDeployAadManifest,
-} from "../../../../src/component/driver/aad/update";
+import { UpdateAadAppDriver } from "../../../../src/component/driver/aad/update";
 import {
   MockedLogProvider,
   MockedM365Provider,
@@ -26,11 +23,7 @@ import {
 } from "../../../../src/component/driver/aad/error/unhandledError";
 import { cwd } from "process";
 import { InvalidActionInputError, UnresolvedPlaceholderError } from "../../../../src/error/common";
-import { Inputs, Platform, v2, ok, err } from "@microsoft/teamsfx-api";
-import os from "os";
-import { MockTools, randomAppName } from "../../../core/utils";
-import { CoreQuestionNames } from "../../../../src/core/question";
-import { environmentManager } from "../../../../src/core/environment";
+
 chai.use(chaiAsPromised);
 const expect = chai.expect;
 
@@ -51,10 +44,6 @@ describe("aadAppUpdate", async () => {
     logProvider: new MockedLogProvider(),
     projectPath: cwd(),
     ui: new MockedUserInteraction(),
-  };
-  const inputs: v2.InputsWithProjectPath = {
-    platform: Platform.VSCode,
-    projectPath: path.join(os.tmpdir(), randomAppName()),
   };
 
   let envRestore: RestoreFn | undefined;
@@ -542,42 +531,5 @@ describe("aadAppUpdate", async () => {
 
     expect(result.result.isErr()).to.be.true;
     expect(result.result._unsafeUnwrapErr()).is.instanceOf(UnresolvedPlaceholderError);
-  });
-
-  it("if getQuestionForDeployAadManifest not dynamic", async () => {
-    inputs.platform = Platform.CLI_HELP;
-    const nodeRes = await getQuestionForDeployAadManifest(inputs);
-    chai.assert.isTrue(nodeRes.isOk() && nodeRes.value == undefined);
-  });
-
-  it("getQuestionForDeployAadManifest happy path", async () => {
-    inputs.platform = Platform.VSCode;
-    inputs[CoreQuestionNames.AadAppManifestFilePath] = "aadAppManifest";
-    inputs[CoreQuestionNames.TargetEnvName] = "dev";
-    sinon.stub(fs, "pathExistsSync").returns(true);
-    sinon.stub(environmentManager, "listAllEnvConfigs").resolves(ok(["dev", "local"]));
-    const nodeRes = await getQuestionForDeployAadManifest(inputs);
-    chai.assert.isTrue(nodeRes.isOk());
-    if (nodeRes.isOk()) {
-      const node = nodeRes.value;
-      chai.assert.isTrue(node != undefined && node?.children?.length == 2);
-      const aadAppManifestQuestion = node?.children?.[0];
-      const envQuestion = node?.children?.[1];
-      chai.assert.isNotNull(aadAppManifestQuestion);
-      chai.assert.isNotNull(envQuestion);
-    }
-  });
-  it("getQuestionForDeployAadManifest without env", async () => {
-    inputs.platform = Platform.VSCode;
-    inputs[CoreQuestionNames.AadAppManifestFilePath] = "aadAppManifest";
-    inputs[CoreQuestionNames.AadAppManifestFilePath] = "aadAppManifest";
-    inputs[CoreQuestionNames.TargetEnvName] = "dev";
-    sinon.stub(fs, "pathExistsSync").returns(false);
-    const nodeRes = await getQuestionForDeployAadManifest(inputs);
-    chai.assert.isTrue(nodeRes.isOk());
-    if (nodeRes.isOk()) {
-      const node = nodeRes.value;
-      chai.assert.isTrue(node != undefined && node?.children?.length == 1);
-    }
   });
 });
