+++ packages/fx-core/tests/core/question.test.ts
@@ -9,7 +9,7 @@ import {
   ProgrammingLanguageQuestion,
   ScratchOptionYesVSC,
 } from "../../src/core/question";
-import { FuncValidation, Inputs, Platform, QTreeNode } from "@microsoft/teamsfx-api";
+import { FuncValidation, Inputs, Platform, QTreeNode, v2, ok, err } from "@microsoft/teamsfx-api";
 import {
   BotNewUIOptionItem,
   BotOptionItem,
@@ -31,7 +31,12 @@ import {
 import { getLocalizedString } from "../../src/common/localizeUtils";
 import { addOfficeAddinQuestions } from "../../src/core/middleware/questionModel";
 import * as featureFlags from "../../src/common/featureFlags";
-
+import os from "os";
+import { MockTools, randomAppName } from "./utils";
+import { CoreQuestionNames, getQuestionForDeployAadManifest } from "../../src/core/question";
+import { environmentManager } from "../../src/core/environment";
+import path from "path";
+import * as fs from "fs-extra";
 describe("Programming Language Questions", async () => {
   it("should return csharp on VS platform", async () => {
     chai.assert.isTrue(ProgrammingLanguageQuestion.dynamicOptions !== undefined);
@@ -324,5 +329,52 @@ describe("addOfficeAddinQuestions()", () => {
       originOption.label,
       `$(new-folder) ${getLocalizedString("core.ScratchOptionYesVSC.label")}`
     );
+  });
+});
+
+describe("updateAadManifestQeustion()", async () => {
+  const inputs: v2.InputsWithProjectPath = {
+    platform: Platform.VSCode,
+    projectPath: path.join(os.tmpdir(), randomAppName()),
+  };
+
+  afterEach(async () => {
+    sinon.restore();
+  });
+  it("if getQuestionForDeployAadManifest not dynamic", async () => {
+    inputs.platform = Platform.CLI_HELP;
+    const nodeRes = await getQuestionForDeployAadManifest(inputs);
+    chai.assert.isTrue(nodeRes.isOk() && nodeRes.value == undefined);
+  });
+
+  it("getQuestionForDeployAadManifest happy path", async () => {
+    inputs.platform = Platform.VSCode;
+    inputs[CoreQuestionNames.AadAppManifestFilePath] = "aadAppManifest";
+    inputs[CoreQuestionNames.TargetEnvName] = "dev";
+    sinon.stub(fs, "pathExistsSync").returns(true);
+    sinon.stub(environmentManager, "listAllEnvConfigs").resolves(ok(["dev", "local"]));
+    const nodeRes = await getQuestionForDeployAadManifest(inputs);
+    chai.assert.isTrue(nodeRes.isOk());
+    if (nodeRes.isOk()) {
+      const node = nodeRes.value;
+      chai.assert.isTrue(node != undefined && node?.children?.length == 2);
+      const aadAppManifestQuestion = node?.children?.[0];
+      const envQuestion = node?.children?.[1];
+      chai.assert.isNotNull(aadAppManifestQuestion);
+      chai.assert.isNotNull(envQuestion);
+    }
+  });
+  it("getQuestionForDeployAadManifest without env", async () => {
+    inputs.platform = Platform.VSCode;
+    inputs[CoreQuestionNames.AadAppManifestFilePath] = "aadAppManifest";
+    inputs[CoreQuestionNames.AadAppManifestFilePath] = "aadAppManifest";
+    inputs[CoreQuestionNames.TargetEnvName] = "dev";
+    sinon.stub(fs, "pathExistsSync").returns(false);
+    const nodeRes = await getQuestionForDeployAadManifest(inputs);
+    chai.assert.isTrue(nodeRes.isOk());
+    if (nodeRes.isOk()) {
+      const node = nodeRes.value;
+      chai.assert.isTrue(node != undefined && node?.children?.length == 1);
+    }
   });
 });
